test: Piece 테스트 추출
