+++ README.md
@@ -39,12 +39,17 @@
     - 체스 (Chess)
 
 - 2단계
+  
+  - TODO : Knight 테스트 케이스 추가!
+
   - 체스 규칙
     - [x] 각 플레이어는 번갈아가며 반드시 기물을 움직여야 함.
     - [x] 움직이려는 기물의 위치에서 기물이 없으면 예외 발생
     - [x] 움직이려는 위치에 같은 색깔의 기물이 있을 경우 예외 발생
     - [x] 나이트를 제외한 모든 기물은 앞에 장애물이 있을 경우 넘어서 이동 불가
     - [x] 나이트는 장애물을 넘어서 이동 가능
+    - [ ] 폰은 이동하려는 위치에 기물이 있는 경우 예외 발생
+    - [ ] 폰을 제외한 다른 색깔의 기물이 위치한 곳으로 이동 시 해당 기물을 잡는다.
     
   - 기물 이동
     - 폰
@@ -53,8 +58,8 @@
       - [x] 후퇴가 불가능
       - [x] 양 옆으로 이동 불가
       - [ ] 상대방의 피스는 대각선 한 칸으로만 잡을 수 있다.
-      - (프로모션) 폰이 상대방 진영 끝까지 갔을때 퀸, 룩, 비숍, 나이트로 바꿀 수 있다. (선택 사항)
-      - (앙파상) 흰색 폰이 5번 랭크에 있고 검은색 폰이 처음 2칸 전진을 하면 앙파상을 이용해 대각선으로 이동하면서 검은색 폰을 넘은 규칙으로 검은색 폰을 잡을 수 있다. (선택 사항)
+      - [ ] (프로모션) 폰이 상대방 진영 끝까지 갔을때 퀸, 룩, 비숍, 나이트로 바꿀 수 있다. (선택 사항)
+      - [ ] (앙파상) 흰색 폰이 5번 랭크에 있고 검은색 폰이 처음 2칸 전진을 하면 앙파상을 이용해 대각선으로 이동하면서 검은색 폰을 넘은 규칙으로 검은색 폰을 잡을 수 있다. (선택 사항)
     - 룩
       - [x] 앞으로 제한 없이 이동 가능
       - [x] 뒤로 제한 없이 이동 가능
@@ -67,7 +72,7 @@
       - (캐슬링) 룩하고 캐슬링이 가능하다. (선택 사항)
     - 나이트
       - [x] 두 칸 전진 + 좌우로 한 칸 움직일 수 있다. (장기의 마처럼 이동)
-      - [ ] 피스들 중에서 유일하게 다른 피스를 뛰어넘을 수 있다.
+      - [x] 피스들 중에서 유일하게 다른 피스를 뛰어넘을 수 있다.
     - 비숍 
       - [x] 대각선 방향으로 칸 수의 제한 없이 움직일 수 있다. 
       - [x] 어두운 칸 비숍, 밝은 칸 비숍은 자기의 색의 칸에서만 이동이 가능하다.
